criando erros costumizado

+++ src/storage/group/groupCreate.ts
@@ -1,4 +1,5 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
+import { AppError } from '@utils/AppError';
 import {GROUP_COLLECTION} from '@storage/storageConfig'
 import { groupsGetAll } from "./groupsGetAll";
 
@@ -6,6 +7,11 @@ export async function  groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
 
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if(groupAlreadyExists) {
+      throw new AppError('Já existe um grupo cadastrado com esse nome.')
+    }
     await AsyncStorage.setItem(GROUP_COLLECTION,JSON.stringify([...storedGroups,newGroup]))
 
   } catch (error) {
